[Refactor] (test) Move the modules which are sample components for integration testing into a single sub-package.
